+++ changelog.md
@@ -2,6 +2,7 @@
 
 * Updated Translations
 * Fixed incompatibility with some mods (including Dungeons Gear)
+* Added support for Mod Menu and Cloth Config
 
 # 1.5.2